+++ models/task/package.json
@@ -35,7 +35,6 @@
     "@anticrm/model-workbench": "~0.6.1",
     "@anticrm/model-contact": "~0.6.1",
     "@anticrm/task": "~0.6.0",
-    "@anticrm/task-resources": "~0.6.0",
     "@anticrm/model-chunter": "~0.6.0",
     "@anticrm/workbench": "~0.6.1"
   }
