+++ plugins/task-resources/src/index.ts
@@ -21,8 +21,6 @@ import CreateProject from './components/CreateProject.svelte'
 import TaskPresenter from './components/TaskPresenter.svelte'
 import KanbanCard from './components/KanbanCard.svelte'
 
-export { createProjectKanban } from './utils'
-
 export default async (): Promise<Resources> => ({
   component: {
     CreateTask,
