+++ plugins/task-resources/src/utils.ts
@@ -43,41 +43,3 @@ export async function uploadFile (space: Ref<Space>, file: File, attachedTo: Ref
   console.log(uuid)
   return uuid
 }
-
-export async function createProjectKanban (
-  projectId: Ref<Project>,
-  factory: <T extends Doc>(_class: Ref<Class<T>>, space: Ref<Space>, data: Data<T>, id: Ref<T>) => Promise<void>
-): Promise<void> {
-  const states = [
-    { color: '#7C6FCD', name: 'Open' },
-    { color: '#6F7BC5', name: 'In Progress' },
-    { color: '#77C07B', name: 'Under review' },
-    { color: '#A5D179', name: 'Done' },
-    { color: '#F28469', name: 'Invalid' }
-  ]
-  const ids: Array<Ref<State>> = []
-  for (const st of states) {
-    const sid = (projectId + '.state.' + st.name.toLowerCase().replace(' ', '_')) as Ref<State>
-    await factory(
-      core.class.State,
-      projectId,
-      {
-        title: st.name,
-        color: st.color
-      },
-      sid
-    )
-    ids.push(sid)
-  }
-
-  await factory(
-    view.class.Kanban,
-    projectId,
-    {
-      attachedTo: projectId,
-      states: ids,
-      order: []
-    },
-    (projectId + '.kanban.') as Ref<Kanban>
-  )
-}
