+++ plugins/task/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@anticrm/platform": "~0.6.5",
     "@anticrm/core": "~0.6.11",
-    "@anticrm/contact": "~0.6.2"
+    "@anticrm/contact": "~0.6.2",
+    "@anticrm/view": "~0.6.0"
   }
 }
